+++ examples/image_classification/image-classification.py
@@ -1,45 +1,50 @@
 import argparse
 import sys
 import logging
+from torch.utils.data import DataLoader
+
+from utils.utils import (
+    create_optimizer,
+    create_scheduler,
+    set_logging,
+)
+from utils.resultlogger import ResultLogger
+from utils.checkpointmanager import CheckpointManager
+from utils.experimentcreator import ExperimentCreator
+from utils.etaestimator import ETAEstimator
+from utils.arg_parser import create_argparser
+from train import train_model
 
 
 sys.path.append("../../")
 
-from arg_parser import create_argparser  # noqa: E402
-from train import train_model  # noqa: E402
 from bitorch.datasets.base import Augmentation  # noqa: E402
 from bitorch.models import model_from_name  # noqa: E402
-from torch.utils.data import DataLoader  # noqa: E402
 from bitorch.datasets import dataset_from_name  # noqa: E402
-# TODO: add tensorboard support!
-# from tensorboardX import SummaryWriter
 
 
-def set_logging(args: argparse.Namespace) -> None:
-    """applies cli configuration to logging
-
-    Args:
-        args (argparse.Namespace): cli arguments for logging containing log_level and log_file. if log_file is none,
-        logging will be printed to stdout.
-    """
-    log_level_name = args.log_level.upper()
-    log_level = getattr(logging, log_level_name)
-
-    if args.log_file is None:
-        logging.basicConfig(format='%(asctime)s - %(levelname)s: %(message)s', level=log_level, force=True)
-    else:
-        logging.basicConfig(filename=args.log_file, format='%(asctime)s - %(levelname)s: %(message)s',
-                            level=logging.DEBUG, force=True)
-
-
-def main(args: argparse.Namespace, model_args: argparse.Namespace) -> None:
-    """trains a model on the configured image dataset
+def main(
+        parser: argparse.ArgumentParser,
+        args: argparse.Namespace,
+        model_parser: argparse.ArgumentParser,
+        model_args: argparse.Namespace) -> None:
+    """trains a model on the configured image dataset.
 
     Args:
+        parser (argparse.ArgumentParser): parser for cli args (used for experiment creation and)
         args (argparse.Namespace): cli arguments
-        model_args (argparse.Namespace): cli model specific arguments
+        model_parser (argparse.ArgumentParser): parser for model cli args (used for experiment creation)
+        model_args (argparse.Namespace): model specific cli arguments
     """
-    set_logging(args)
+    set_logging(args.log_file, args.log_level, args.log_stdout)
+    if args.experiment:
+        experimentCreator = ExperimentCreator(args.experiment_name, args.experiment_dir, __file__)
+        experimentCreator.create(parser, args, model_parser, model_args)
+        experimentCreator.run_experiment()
+
+    result_logger = ResultLogger(args.result_file, args.tensorboard, args.tensorboard_output)
+    checkpoint_manager = CheckpointManager(args.checkpoint_dir, args.checkpoint_keep_count)
+    eta_estimator = ETAEstimator(args.eta_file, args.log_interval)
 
     dataset = dataset_from_name(args.dataset)
     if dataset.name == 'imagenet' and args.nv_dali:
@@ -58,16 +63,25 @@ def main(args: argparse.Namespace, model_args: argparse.Namespace) -> None:
         test_loader = DataLoader(test_dataset, batch_size=args.batch_size, num_workers=args.num_workers)
 
     model_arg_dict = vars(model_args)
-    logging.info(f"got model args as dict: {model_arg_dict}")
+    logging.debug(f"got model args as dict: {model_arg_dict}")
 
     model = model_from_name(args.model)(**model_arg_dict, dataset=dataset)  # type: ignore
     logging.info(f"using {model.name} model...")
 
+    optimizer = create_optimizer(args.optimizer, model, args.lr, args.momentum)
+    scheduler = create_scheduler(args.lr_scheduler, optimizer, args.lr_factor,
+                                 args.lr_steps, args.epochs)  # type: ignore
+
+    if args.checkpoint_load:
+        model, optimizer, scheduler, start_epoch = checkpoint_manager.load_checkpoint(
+            args.checkpoint_load, model, optimizer, scheduler, args.pretrained)
+    else:
+        start_epoch = 0
+
     gpus = False if args.cpu or not args.gpus else ','.join(args.gpus)
-    train_model(model, train_loader, test_loader, epochs=args.epochs, optimizer_name=args.optimizer,
-                lr_scheduler=args.lr_scheduler, lr_factor=args.lr_factor, lr_steps=args.lr_steps,
-                momentum=args.momentum,
-                lr=args.lr, log_interval=args.log_interval, gpus=gpus)
+    train_model(model, train_loader, test_loader, start_epochs=start_epoch, epochs=args.epochs, optimizer=optimizer,
+                scheduler=scheduler, lr=args.lr, log_interval=args.log_interval, gpus=gpus,
+                result_logger=result_logger, checkpoint_manager=checkpoint_manager, eta_estimator=eta_estimator)
 
 
 if __name__ == "__main__":
@@ -75,4 +89,4 @@ if __name__ == "__main__":
     args, unparsed_model_args = parser.parse_known_args()
     model_args = model_parser.parse_args(unparsed_model_args)
 
-    main(args, model_args)
+    main(parser, args, model_parser, model_args)