+++ CHANGELOG.md
@@ -20,3 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
     - swishsign
 - added support for cifar10 and mnist
 - adds general training script for image classification
+- result logger
+- checkpoint manager
+- eta estimator
+- experiment creator